Move k-means implementation from diskann-providers to diskann-disk (#933)
K-means in `diskann-providers` was the last consumer of the old
BLAS-based clustering path; PQ training has since migrated to
`diskann-quantization`. The only active call site remaining was
disk-index partitioning in `diskann-disk`.

We will keep diskann-providers's implementation for now and move it to
`diskann-disk`, rather than switching to the one in
diskann-quantization, for the following reasons:
- K-means in diskann-providers performs better at higher dimensions
(>100):
<img width="618" height="507" alt="image"
src="https://github.com/user-attachments/assets/1e483411-18ae-4cc7-aa59-d9df05f4e0cf"
/>

- K-means in diskann-providers supports multi-threading:
<img width="612" height="503" alt="image"
src="https://github.com/user-attachments/assets/f5219632-6223-45fe-b0e0-18d40f0e2a1d"
/>

We will work on closing these performance gaps and converging the two
implementations in separate PRs.

# Changes in this PR

## diskann-disk
- Added `src/utils/kmeans.rs` — k-means implementation moved from
`diskann-providers`
- Added `src/utils/math_util.rs` — mathematical utilities
(`compute_vecs_l2sq`, `compute_closest_centers`,
`compute_closest_centers_in_block`, and helpers) extracted from
`diskann-providers` and deduplicated
- Exported `k_means_clustering`, `k_meanspp_selecting_pivots`,
`run_lloyds`, `compute_vecs_l2sq`, `compute_closest_centers`,
`compute_closest_centers_in_block` from `utils/mod.rs`
- Updated `utils/partition.rs` to import kmeans functions and math
utilities from local modules instead of `diskann-providers`
- Moved kmeans criterion and iai-callgrind benchmarks from
`diskann-providers/benches` to `diskann-disk/benches`
- Added `proptest` and `approx` to dev-dependencies

## diskann-providers
- Deleted `src/utils/kmeans.rs`
- Removed `k_means_clustering`, `k_meanspp_selecting_pivots`,
`run_lloyds`, `compute_vecs_l2sq`, `compute_vec_l2sq` from the public
API
- Removed the now-deduplicated math utility implementations from
`math_util.rs`
- Removed dead OPQ code: `generate_optimized_pq_pivots`,
`opq_quantize_all_chunks`, `copy_chunk_centroids_to_full_table`, their
test, and unused imports/constants — these were the sole remaining
callers of k-means in this crate and were already gated behind
`#[allow(dead_code)]`

---------
