+++ diskann-disk/benches/benchmarks_iai/mod.rs
@@ -4,3 +4,4 @@
  */
 
 pub mod aligned_file_reader_bench_iai;
+pub mod kmeans_bench_iai;
