+++ diskann-disk/src/utils/mod.rs
@@ -11,6 +11,12 @@
 pub mod partition;
 pub use partition::partition_with_ram_budget;
 
+pub mod math_util;
+pub use math_util::{compute_closest_centers, compute_closest_centers_in_block, compute_vecs_l2sq};
+
+pub mod kmeans;
+pub use kmeans::{k_means_clustering, k_meanspp_selecting_pivots, run_lloyds};
+
 pub mod instrumentation;
 
 pub mod aligned_file_reader;
