+++ diskann-providers/benches/bench_main.rs
@@ -8,7 +8,6 @@ use benchmarks::{
     compute_pq_bench::benchmark_compute_pq,
     copy_aligned_data_bench::benchmark_copy_aligned_data,
     diskann_bench::benchmark_diskann_insert,
-    kmeans_bench::benchmark_kmeans,
     neighbor_bench::{
         benchmark_priority_queue_has_notvisited_node, benchmark_priority_queue_insert,
     },
@@ -18,7 +17,6 @@ mod benchmarks;
 
 criterion_group!(
     benches,
-    benchmark_kmeans,
     benchmark_priority_queue_insert,
     benchmark_compute_pq,
     benchmark_diskann_insert,
