+++ diskann-providers/benches/bench_main_iai.rs
@@ -7,8 +7,7 @@ use benchmarks_iai::{
     chunking_closest_centers_benchmark_iai::benchmark_chunking_size_closest_centers_performance_bench_iai,
     compute_pq_iai::compute_pq_bench_iai,
     copy_aligned_data_bench_iai::benchmark_copy_aligned_data_bench_iai,
-    diskann_iai::diskann_insert_bench_iai, kmeans_bench_iai::kmeans_bench_iai,
-    neighbor_bench_iai::priority_queue_insert_bench_iai,
+    diskann_iai::diskann_insert_bench_iai, neighbor_bench_iai::priority_queue_insert_bench_iai,
 };
 use iai_callgrind::{EventKind, LibraryBenchmarkConfig, RegressionConfig, main};
 mod benchmarks_iai;
@@ -23,7 +22,6 @@ main!(
         benchmark_chunking_size_closest_centers_performance_bench_iai,
         benchmark_copy_aligned_data_bench_iai,
         compute_pq_bench_iai,
-        kmeans_bench_iai,
         priority_queue_insert_bench_iai,
         diskann_insert_bench_iai,
 );
