+++ diskann-providers/benches/benchmarks/mod.rs
@@ -7,5 +7,4 @@ pub mod chunking_closest_centers_benchmark;
 pub mod compute_pq_bench;
 pub mod copy_aligned_data_bench;
 pub mod diskann_bench;
-pub mod kmeans_bench;
 pub mod neighbor_bench;
