+++ diskann-providers/benches/benchmarks_iai/mod.rs
@@ -7,5 +7,4 @@ pub mod chunking_closest_centers_benchmark_iai;
 pub mod compute_pq_iai;
 pub mod copy_aligned_data_bench_iai;
 pub mod diskann_iai;
-pub mod kmeans_bench_iai;
 pub mod neighbor_bench_iai;
